designed current weather ui

+++ app/src/main/java/com/sunday/sustainweatherforcast/data/db/unitLocalized/MetricCurrentWeatherEntry.kt
@@ -2,16 +2,16 @@ package com.sunday.sustainweatherforcast.data.db.unitLocalized
 
 data class MetricCurrentWeatherEntry (
 
-    override val feelslike: Int,
+    override val feelslike: Double,
     override val isDay: String,
-    override val precip: Int,
-    override val temperature: Int,
-    override val visibility: Int,
-    override val weatherCode: Int,
+    override val precip: Double,
+    override val temperature: Double,
+    override val visibility: Double,
+    override val weatherCode: Double,
     override val weatherDescriptions: List<String>,
     override val weatherIcons: List<String>,
     override val windDir: String,
-    override val windSpeed: Int
+    override val windSpeed: Double
 
 
 ): UnitSpecificCurrentWeatherEntry
