+++ app/src/main/java/com/sunday/sustainweatherforcast/data/db/unitLocalized/UnitSpecificCurrentWeatherEntry.kt
@@ -2,14 +2,14 @@ package com.sunday.sustainweatherforcast.data.db.unitLocalized
 
 
 interface UnitSpecificCurrentWeatherEntry {
-    val feelslike: Int
+    val feelslike: Double
     val isDay: String
-    val precip: Int
-    val temperature: Int
-    val visibility: Int
-    val weatherCode: Int
+    val precip: Double
+    val temperature: Double
+    val visibility: Double
+    val weatherCode: Double
     val weatherDescriptions: List<String>
     val weatherIcons: List<String>
     val windDir: String
-    val windSpeed: Int
+    val windSpeed: Double
 }
